[mojo-lang][CheckLifetimes] Fix hlcf.if result processing.
This fixes CheckLifetimes' handling of hlcf.if ops when they have
owned non-trivial results. This can occur in and/or expressions with
non-trivial condition types (as happens with PythonObject).  We were
processing the def of the 'if' after the then/else arm instead of
processing it at the bottom of the 'if', leading to propagating liveness
into the arms where the value doesn't even dominate.

This fixes https://github.com/modular/modular/issues/4694

MODULAR_ORIG_COMMIT_REV_ID: 5f684c0425a422f3b0f9bb497db33c15f07ca43d