[mojo-lang] Implement support for list "for" comprehensions.
This implements support for list "comprehensions" like
`[x*y for x in range(42)]`.  This only supports lists and only
supports 'for' (not 'if') so far, but is a big step forward in terms
of all the infra.  More details in the design doc.

MODULAR_ORIG_COMMIT_REV_ID: be1427f0e4df3bff522b1b802af50c8378c33992