[mojo-lang] Add infra to resolve LHS from RHS in an `=` expr.
This introduces infra in the compiler to allow it to infer a RHS
type from a LHS when handling an "x = []" sort of expression.  We
need to (and could already) infer the LHS type from the RHS (which
is needed when implicitly declaring variables in a def) but we also
need to do the other thing when the LHS has a known type.

This adds plumbing to the compiler so we can speculatively emit an
expression that is about to be assigned to.  It can either fail,
produce an LValue or have an unresolvable type.

Building on that infra, we teach DeclRefNode to participate, so we
get a few cases to handle this.  This isn't enough to resolve the
fullness of https://github.com/modular/modular/issues/4684 or MOCO-1972
because
we don't support `x.y` expressions yet.

MODULAR_ORIG_COMMIT_REV_ID: fba289d270de803e967fc85a1d721920c4a80447