[Mojo] Extend the Symbol Array To Include Triple (#61920)
We need to lower the lit representation of a closure to the kgen
representation of a closure. Before this patch the lit representation
stored an attribute per capture. This attribute was either a symbol or
an origin to indicate capture by copy/move or reference. This patch
changes the symbol to be a mem triple attribute, rendering it compatible
with its kgen version so we can lower it to kgen in a follow up PR.

MODULAR_ORIG_COMMIT_REV_ID: 9c57598df090068ed88838b03ba60f3373967137