[mojo-lang] Improve for loop indvar emission a bit (#62616)
This changes the emission of for loop statements to emit the vardecl
for the indvar into the body of the loop (better scoping the SSA value
and simplifying the logic) instead of emitting it outside the loop.

This also changes the emission of "as" expressions in "with" statements
to be more consistent with other target-producing things.

MODULAR_ORIG_COMMIT_REV_ID: 40b491ab35583f10a5c2e1287cd2496cd00626d8