[mojo-lang] Switch for stmts + with blocks to use target emission. (#62707)
This enhances target emission to support "var" style targets, which
create a vardecl in the current scope instead of at function scope.

This further uses this to emit the target expression of a 'for' and
'with' statement with consolidated logic.  This allows access to the
broader grammar of targets, which include parens, subscripts, and
tuple expressions.

MODULAR_ORIG_COMMIT_REV_ID: f85942d36ded19987820d11a2fb64fa854c518d9