[MOGG] Incorporate KGEN module hash into caching strategy (#62723)
The current structure of the MOGG cache does not currently factor in the
state of the KGEN module derived from the MojoLibraryAnalysis. This KGEN
module contains the implementations of all the backing kernels
referenced during the stub compilation in JitCompileKernels.

When one of the backing kernels is modified, that modification is not
reflected in the hash used to look up if a particular KGEN generator has
been compiled previously, which results in erroneous cache hits. This is
because the caching strategy hashes the KGEN generator, but not any of
its dependencies invoked through `kgen.call`.

This solution adds the hash of the KGEN module from MojoLibraryAnalysis
to the 'environment' has when resolving the corresponding MOGG cache.
This is not particularly granular, but allows us to continue treating
the MLIR operations as opaque blobs hashed via their bytecode
representation.

Closes: GEX-2216
Closes: GEX-2291
MODULAR_ORIG_COMMIT_REV_ID: adea1a1ee01ca6ad97009c29cd53b53bbaec8dea