[Serve] Remove StaticZMQ dispatcher transport
This PR removes the StaticZMQ dispatcher transport. This was primarily
added for testing but, it did not fit well with REQ/RESP pattern of the
dispatcher as it is hard to distinguish what is a request and what is a
response. Hence, instead of making it work, this PR removes it. For
testing, `default_destination_address` option is added to DynamicZMQ.
This can set as follows:

```
MAX_SERVE_DISPATCHER_CONFIG='{"transport": "dynamic_zmq", "transport_config": {"bind_address": "tcp://localhost:5555", "default_destination_address": "tcp://localhost:5556"}}'
```

When this is set, if not destination is provided for send operations,
this default destination is automatically used.

Lastly, the PR also adds more tests and fixes some bugs.

MODULAR_ORIG_COMMIT_REV_ID: 3631540a180c264a25b38fdffaec5440fc6e37fa