[mojo-lang] Refactor `except` parsing a bit. (#62819)
This refactors except parsing to be closer to right, but (as noted
in the fixme) we are just completely wrong at the moment.  This does
fix the bug, but simplifies the code a bit. I don't plan to work
on this in the near future.

MODULAR_ORIG_COMMIT_REV_ID: 3a707ed86cc4265ad7e75309972cb537ba73c37d