[stdlib] [proposal] Binary IO API
This proposal covers binary IO via an "IO Engine" concept, designed to
be highly modular and composable while allowing a mix of state of the
art/near future IO APIs with older, more familiar APIs. Although this
proposal does provide example APIs, an additional aspect up for
discussion is whether this design philosophy is desirable. The primary
goal is to allow Mojo to adopt new IO APIs as they are developed,
instead of providing a central abstraction which Mojo is chained to (see
C++ iostreams), while still allowing abstraction over commonalities.
This is designed to function even in the cases where some capabilities
are absent, such as GPUs often lacking access to a full network stack. It does
somewhat increase the implementation burden of the standard library, but
does so for what I think are justifiable reasons.

Signed-off-by: Owen Hilyard <hilyard.owen@gmail.com>