[mojo-repl] Allow uninitialized values live-out of cells. (#62841)
The mojo repl doesn't have any lifetime analysis and doesn't work
with CheckLifetimes, and persists values across cells.

This patch adds a very unprincipled hack, allow uninitialized
values at the end of REPL cells. The reason we need this is that
the REPL "persists" values onto the heap with an IR rewrite, and
does so before lifetime checking.  As such, it has no idea what
values are live out of the end of each cell.  This needs to be
 fixed, but allow simple things like integers to "work" for now like
this in the repl:

  `var even = [i for i in range(5) if i % 2 == 0]`

This will not work at all for non-trivial values though because
reassignments over them will assume they are initialized.

MODULAR_ORIG_COMMIT_REV_ID: 63d49d2fb9ae62d6ad32d4816b21dc0df7cc2612