[mojo-lang] Fix thorny debug info problem with comprehensions.
Mojo can't know the type of the ultimate collection until it emits a
pile of for loops and if statements so it can infer the element type
that is generated, which then influences the ultimate collection.
It needs to thus generate the type for the ultimate var decl in a
deeply nested pile of scopes, but that decl is outside of the scopes.

We were previously not accounting for this in debug info generation:
generating the synthesized var decl with the debug scope within
the for loops, even though the decl was outside of it.  Adjust it back
to the right debug scope to avoid a crash with debug info on.

MODULAR_ORIG_COMMIT_REV_ID: f1fc9c1f0f6335e555620800d3b8a081d27ae658