[mojo-lang] Fix a.x lookup when 'a' is a package or file. (#62878)
This fixes a crash and other bad behavior when doing qualified
lookup in a package or file.  Instead of switching the context of the
expr emitter, we emit the expr into the current context, but do the
lookup in the package scope.

This fixes MOCO-2006

MODULAR_ORIG_COMMIT_REV_ID: 9d44bf2d448d75133ef7a9fa459c83188d6e1274