[MOGG] Improvements for getBytecodeHash (#62912)
The getBytecodeHash function is used internally in several places in the
codebase:

- HoistHarness: for uniquing hoisted stub functions
- MOGGCompiler: for name mangling to create unique names for stub
functions
- MOGG Cache: for generating the keys to the MOGG cache, this needs to
  hash both the kernel and the Mojo code from MojoLibraryAnalysis.

Now that the code from MojoLibraryAnalysis is factored into the hash of
the MOGG cache, getBytecodeHash is now processing a large amount of IR.
The current implementation uses a very fast hash function, but has to do
a large amount of pre-processing to get the desired behavior:

1. Clone the entire MLIR operation being hashed (which could be the
   entire library of MOGG kernels).
2. Strip all the location information from the cloned module.
3. Serialize the IR to a bytecode buffer.
4. Hash the bytecode buffer.

The first 3 step can be elided by using a hash operation that supports
incremental updates and wrapping that as a raw_ostream for the bytecode
writer. This change avoids the allocations by using the streaming
variants of xxhash from the standalone xxhash library.

MODULAR_ORIG_COMMIT_REV_ID: df161498a1147bd15f8aaf637c0a326ddb5295f3