[External] [Mojo-stdlib] Remove `FLAG_IS_STATIC_CONSTANT` from the `String` struct (#62316)
[External] [Mojo-stdlib] Remove `FLAG_IS_STATIC_CONSTANT` from the
`String` struct

While poking around in the `String` struct, I noticed that it was
actually possible to know if we point to a `StaticString` without the
need to read the flag `FLAG_IS_STATIC_CONSTANT`. We can use the
condition "pointer is not null and capacity is 0".
I also simplify the `String()` constructor in the process. Let's remove
the flag and make it available for future use cases.

See
[Internal link]
for more info on the subject:
> In addition to setting the bit to know whether the pointer is to a
constant string, it is important to know that the String type also sets
its notion of "capacity" to zero. This ensures that any attempt to
append or access will immediately reallocate without extra checks. This
is a minor optimization and simplification of code, but it means that
static constant strings will have capacity=0 but have length=n where n >
0

@lattner expressed his desire to do this in [this discord
message](https://discordapp.com/channels/1087530497313357884/1374762032254943284/1374884885809856642)

Co-authored-by: gabrieldemarmiesse <gabrieldemarmiesse@gmail.com>
Closes modular/modular#4661
MODULAR_ORIG_COMMIT_REV_ID: 705d5ab00efc5b67ecabd7987cf33a05be6b3411