[mojo-lang] Change `for` bindings to being like `read` args.
This patch changes bindings without a pattern specifier in a `for`
and or `with` statement like `for a, b in ...` to behave like a `read`
argument binding.

Specifically, the bound name acts like an immutable reference in the
body of the loop. It is formed either by binding to the reference
returned by a `ref`-yielding iterator (and stripping it's mutability) or
by storing the value returned by a value-returning iterator into a local
variable and using it.

This gives a predictable and easy to understand model, which gives
high performance (never copying) automatically.  When you want a
locally mutable copy of a value, you can use `var` and when you want
to bind to a mutable reference, use `ref`.  Easy peasy.

Thank you to the Mojo community for feedback on the proposal. This
updates the doc to follow the new design, and unrelatedly updates the
"fixing simple literals" doc to use init syntax.

MODULAR_ORIG_COMMIT_REV_ID: 6f8865318da7e44b06e306466f740e14e0ed434d