[CI] Temporarily pin numpy to fix a float formatting test
builtin/test_float_format.mojo relies on numpy as a reference for
float32 comparisons, but the newly release 2.3.0 changed some formatting
so it is no longer a reliable reference point. As a temporary
workaround, pin to <2.3.0 in these tests specifically.

This isn't an issue in the Bazel build, as that already pins the
version.

MODULAR_ORIG_COMMIT_REV_ID: 15a501c906c14233202df6868b23bc4bbb6885b1