[MO] Add parameters field to mo.opaque (#62998)
This change repurposes the unused metadata field as the place to store
Mojo parameters for the corresponding Mojo type. The parameters are
stored on a DictionaryAttr which naturally provides the normalization
properties that we want.

Towards: GEX-2330
MAX_GRAPH_API_ORIG_REV_ID: fa739d28857cf772d3b1502761454e677994586b