[KGEN] Avoid converting MulNuw to Mul in addition simplification (#63404)
The decomposeAddend canonicalizatio which attempts to unify sum terms
with the same constant multiplier (a * 2 + a * 3) => a * 5 converts all
the POC::MulNuw expressons in the term to the more strict POC::Mul
opcode. This normalization unifes more terms, but can prevent other
unifications in the graph compiler which get to the same a * 5 term by
a different sequence of calculations.

This change has the addend decomposition tracks which kind of
multiplication was used and avoids converting MulNuw to Mul unless at
least one term in the summation already uses the Mul opcode.

Closes: GEX-2342
MODULAR_ORIG_COMMIT_REV_ID: 15270c511b5b0bc333967d1aee3e2b911b7622e0