[mojo] Remove "already inherited from" warning (#63504)
Remove the warning when a later trait is already specified (directly, or
indirectly) by a previous trait in a struct conformance list.

e.g. Previously, this code would emit a warning about struct `Impl`
already inherits from `Parent` via `Child`.
```
trait Parent: pass
trait Child(Parent): pass
struct Impl(Child, Parent): pass
```

The rationale to remove this is that the compiler doesn't really care if
a particular trait is repeated, and sometimes the user does want to
specify a parent trait explicitly for clarity. The current warning is
also problematic because it does not work the same way for trait
compositions, and it causes the ordering of the conformance list to make
a difference in the compilation process (w.r.t. warnings). This latter
issue makes it hard to use a formatter on the conformance list. Any
duplication issue should really be addressed by the linter / formatter.

MODULAR_ORIG_COMMIT_REV_ID: 55b522f87ad4e145065110db24978aa855c8366c