[AsyncRT] Clean up nixl cuda context (#63723)
NIXL CUDA context management

nixlUcxEngine calls vramInitCtx() to create an empty nixlUcxCudaCtx,
which
is a CUcontext and device id tuple. Both are null.

Then we start a progress thread, which calls vramApplyCtx() to set the
context
as active. This is a no-op the first time around.

vramUpdateCtx is called in registerMem. It calls cudaUpdateCtxPtr,

This is checking to see if the pointer we are registering is associated
with
the same device ID and device context as stored in our nixlUcxCudaCtx.
It calls
cuPointerGetAttributes to do this.

If our context was unset, then we set the fields from this call and set
was_updated as an out parameter. This triggers restart of the progress
thread,
so that it can set the right device context thist time.

If our context was set, and it doesn't match, then we return
NIXL_ERR_NOT_SUPPORTED.

Then vramFiniCtx deletes the context in the destructor.

So in summary: nixlUcxEngine only supports one CUDA context. It uses
whatever
context was set on the first call to registerMem. Any other
configuration is
not supported. And we can rip all of this out if we just do context
management
ourselves.

Additionally, the context was only being applied to the progress thread,
but
we need to set it elsewhere. Will do in a followup.

---------

Co-authored-by: Brian <bez@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 8df558d0f379bd3542843225e710beacda48b60b