[mojo-lang] Introduce `kgen.param.for.goto.else` op, ~NFC (#63793)
This introduces a new `kgen.param.for.goto.else` operation that will
allow us to make the 'break' from the body of a param for out to the
'else' branch explicit.  For loops in python are supposed to skip an
'else' when a break or return are used, and we're getting this wrong
(along with several other problems in the neighborhood). This
introduces the infra to solve some of these problems, but doesn't
adopt it yet.

MODULAR_ORIG_COMMIT_REV_ID: 65500ff87792316b8f350ba3bce2f565bf59ab27