[Kernels][GPU] Add a try/catch when trying to find the reduced_batchsize for irfft
The check_error call would throw, but we do not want to exit the
function when trying to find the reduced batch size. Instead, we want to
try getting the estimate and if that fails we want to move on to the
next one.

MODULAR_ORIG_COMMIT_REV_ID: 6c43b73cdd07189dd5e21f5eee3342acaef2ec12