[SDK] Use GPU when available for PyTorch tests (#64318)
All the existing PyTorch integration tests use a CPU target, even if
a GPU is available on the testing machine. This results in the PyTorch
workflow being untested on GPU.

This change updates the tests to detect the availability of a GPU and
execute on the GPU when possible.

MAX_TESTS_TESTS_ORIG_REV_ID: 4498a1d7a884f0614482129b08d4598fc5a5627a