[KGEN] Non-yielding debuginfo concretization (#64309)
This is a problem with our current debuginfo concretization procedure.
Usually during concretization of an attribute, if the current ImplNode
needs to take a blocking dependency on another node, we ask the other
node to reschedule our current node when it completes, and then return
null attribute/type inside the IREvaluator. What is supposed to happen
is that no more concretization happens, and we exit the current task
directly. However, this was not respected during debuginfo
concretization. We didn't handle it correctly when the returned location
was empty, and instead went directly to concretizing the next location.
This lead to us potentially setting multiple "wakers" for the current
ImplNode. This will come back to bite us when eventually multiple
threads start working on the same ImplNode, leading to all kinds of
problems.

This current solution is a short-term stopgap that just avoids us from
ever setting ourselves as waiters of other nodes during debuginfo
concretization. This prevents multiple threads from working on the same
node down the line. This is necessary because unfortunately we don't
have everything set up so that we can save state during debuginfo
concretization, and resume later when another node is done. I created
MOCO-2105 to enable this save & resume behavior for debuginfo.

MODULAR_ORIG_COMMIT_REV_ID: 6df95e6d5a60c51eb21ac8f6d7dc783d63272f66