[TTS] Warn against duplicate request ids
Properly handling this edge case is nontrivial and requires some more
significant changes.

Instead, I just do a best effort to catch such a case and warn the user.

The requests with the duplicate ids may be messed up, but the other
requests are fine.

This is because the exception is isolated to the specific asyncio tasks
and doesn't kill process.

MODULAR_ORIG_COMMIT_REV_ID: 50bda17366ae389e1ea0080ef2551839804ea8dc