[mojo-lang] Fix memory object emission for complex interpreter objects
We noticed that Mojo would sometimes memory when comptime values
get mapped into runtime values. It turns out that this is a 2yo bug (!)
that
has to do with how the comptime interpreter builds a unified memory
model
for a complex series of function calls, e.g. in the testcase, the result
of two
String.add calls that need to heap allocate temporary memory for the
result.

In the comptime interpreter, it needs to model these memory objects, but
it ignores the "free" of them because the old pointer is still
(logically) valid.
Even older logic tried to cache and reuse previously-emitted
materialized
values and tries to do so for a bulk emission of the entire memory
space.

This isn't correct, because this would cause the "free'd" memory objects
to
get materialized with a malloc, but of course the rest of the compiler
has no
idea that these things need to get free'd. Solve this by removing the
caching
(which is no longer needed due to other changes) and simplifying all the
code.

Update memory blob materialization to only materialize used blobs and
their
pointer region blobs recursively to avoid materialize intermediate
memory
generated by the interpreter.

MODULAR_ORIG_COMMIT_REV_ID: 79417ac02049344cf4984ea83defdb4326c4946d