[mojo-stdlib] polish: Rename `IteratorTrait` to `Iterator`
This makes the Mojo trait for modeling a type that provides a
`__next__()` method match the naming of the related type annotation
commonly used in Python

(https://docs.python.org/3/library/collections.abc.html#collections.abc.Iterator).

This also avoids the awkward `*Trait` suffix.

Note that there is a tradeoff here, related to the `Iterable` trait we
will likely add soon. Following standard Mojo naming conventions, the
natural definition of `Iterable` is:

    trait Iterable:
        alias Iterator: Iterator

        fn __iter__(self) -> Iterator:
            ...

Which causes a naming ambiguity in the `alias Iterator: Iterator`.

If we had kept the name `IteratorTrait`, that ambiguity wouldn't exist,
because the alias would be:

        alias Iterator: IteratorTrait

However, with this rename of `IteratorTrait` to `Iterator`, we instead
plan to resolve that alias naming ambiguity using:

        alias IteratorType: Iterator

This is a straightforward tradeoff. Our feeling is that it's better to
use the name `Iterator` for the trait, and push the naming
disambiguating awkwardness down into the associated type.

MODULAR_ORIG_COMMIT_REV_ID: ca30acbf874664f49700e22fa0c1d359df03f9aa