[mojo-parser] Handle implicit conversion of generator type bodies (#64504)
We can allow converting from one GeneratorType to another GeneratorType
if the body of the former generator is implicitly convertible to the
body type of the latter. This only works if we know the body of the
former generator, which means it has to be a PValue, and has to be a
GeneratorAttr so that we can peel out the body. This should be allowed
to relax a lot after the witness table refactor turns metatype casting
into a zero-cost conversion.

MODULAR_ORIG_COMMIT_REV_ID: 98637cffdbe238c169cf5feb77de02daaf9fb4e6