[mojo-lang] Rewrite `kgen.param.for` processing.
This redesigns the implementation of "parameter for" to be more similar
to for loops, using a single body block that contains a "parameter if"
for
the the "execute" and "exit" path instead of having a critical edge
between
the exit and entry of the loop.

This is a pretty big change, but solves a number of problems, including
infinite loops in check lifetimes when trying to insert destructors on
the
critical edge (which was impossible), this fixes
https://github.com/modular/modular/issues/4566?reload=1
and fixes the "else" logic to only be run when it is supposed to, when
the loop exits normally. This fixes
https://github.com/modular/modular/issues/4566?reload=1

MODULAR_ORIG_COMMIT_REV_ID: 7cac693499963600f3aaa9411b3a695f5b4f0fae