[max-examples] fix: Fix constraint failure due to GPU call in histogram.py example
This example was emitting a call to a function that is only supported on
GPU, even when no GPU was available. This changes an `if` to an
`@parameter if` to ensure that the GPU call is statically not emitted if
a GPU is not being targetted.

MODULAR_ORIG_COMMIT_REV_ID: db8b305a4e28c039edb1c64657346c0b5dc17d42