[UCX] Add CUDA archives to `ucx_prebuilt.BUILD`
In [Internal link] we built UCX with
CUDA and uploaded the tarball. In this PR, we update the
`ucx_prebuilt.BUILD` to include the CUDA archives.

Originally this change caused a linker error where the CUDA symbols were
not found. Keith suggested ignoring the linker error and YOLO'ing it at
runtime.

After getting past the linker error, we saw a undefined symbol error at
test runtime for CUDA and NVML symbols. To fix this, we had to make
CUDADriver.cpp load `libcuda.so.1` with `RTLD_GLOBAL` and also
additionally load "libnvidia-ml.so.1". Note that the CUDADriver.cpp
loads NVML but doesn't actually use them. This is just for NIXL.
Now the test passes! However, it is a bit ugly as the UCX CUDA plugin
requires that `libcuda.so.1` and `libnvidia-ml.so.1` are available when
it is being loaded. The desired behavior is that it automatically loads
those libs if they are not available.

To force those libs to be loaded at plugin load time, we instantiate a
unused CUDA DeviceContext at start of the test.

MODULAR_ORIG_COMMIT_REV_ID: ce7f07c2237682989617ab919b2cf4dc72604e2e