[Mojo] Add Missing Argument Conventions and Return Types (#64508)
The signatures of the implementations of move and del for closures and
the synthesis of the symbols of its captures omitted the argument
conventions and return types. As a result, the signature of the symbol
generated by the parser did not match the signature of the symbol
generated by the outline closures pass. This patch corrects that behavior,
by adding argument conventions and result types to the signatures of
the synthesized move and del closure methods.

MODULAR_ORIG_COMMIT_REV_ID: 1f3312a27d4b3c284e7c1190677f9e00d2c0ca69