[SDK] Add a __hash__ method to type and attribute bindings (#64643)
MLIR types and attributes were not providing a hashing method, which
causes them to use the object identity of the binding object. This
results in us being unable to use these types for keys/values in a
semantically relevant way.

This change exposes the internal MLIR hashing operation for types and
attributes, which leverage the identity of the underlying IR object
interned on the MLIRContext.

Closes: GEX-2387
MAX_PYTHON_ORIG_REV_ID: 4e29547c8ea4198d98d1dc65bb147b033fb5058e