[AsyncRT][NIXL][UCX] Ensure DeviceContext is set when performing UCX operations
As UCX calls cuda apis, it expects a active CUDA context to be set. This
means
that we should be within a `DeviceContextScope`. By wrapping methods in
NIXL
UCX backend in a `DeviceContextScope` via `activateContext()`, we are
able
to eliminate the need for `DeviceContextScope ctxScope(*ctx)` in the
unit
test.

~~Additionally, I modify the plugin loading logic to ensure that a CUDA
DeviceContext is prior to dlopen'ing the plugin. This is to ensure that
the relevant libcuda / linvidia-ml symbols are present at plugin load
time.
By doing this this hack here, I can omit this hack in several places
higher
in the stack.~~

~~Now the NixlTestUcx logic is less brittle and will not break if we
create
a NIXL agent prior to creating a DeviceContext like before.~~

The above causes a `malloc_consolidate(): unaligned fastbin chunk
detected
error` when I add "//AsyncRT:DeviceContextImpl" /
"//AsyncRT:DeviceContext"
to the NIXL target's deps. I'm not sure why this is happening. Will
leave
it to another PR.

MODULAR_ORIG_COMMIT_REV_ID: 8adda4535cd7eb48d46e0adedea349a8cc8ffee8