[mojo-lang] Accept `var` as an argument convention.
The `owned` argument convention is the last holdout from the
argument convention renamification.  It has felt odd and people
have found it confusing for a long time, and many folks on the
forum have observed that the existing keyword `var` reflects its
behavior better: this is a named, scoped, mutable owned value,
just like local `var`s and struct fields.

This patch changes the compiler to allow the use of `var` in
argument lists. In a subsequent release we can warn about the use
of `owned`, and then in an even-later release we can remove it.

MODULAR_ORIG_COMMIT_REV_ID: 56114bc8e2e26651e4d1010f6f5f4e21959ae04d