[External] [stdlib] Adds hasher parameter to Dict (#64564)
[External] [stdlib] Adds hasher parameter to Dict

This PR concludes the transformation of the hashing strategy to data
flow based hashing. With this PR users are able to provide a parameter
to the Dict in order to customise the way a hash value is computed for
the keys.

Important note! As at this point in time the AHasher implementation does
not work at compile time and the Dict parameter defaults to AHasher, the
users would need to provide a different hasher (e.g. Fnv1a) if they
chose to have a comp time Dict, Set, or other Dict based data structure.

Co-authored-by: Maxim Zaks <maxim.zaks@gmail.com>
Closes modular/modular#4922
MODULAR_ORIG_COMMIT_REV_ID: c0b8429f7fd9859a2f220d22c26ed4d8d7d14056