[Models] Print both beginning and end of long prompt (#64985)
Previously, the `run_text_generation` methods would only print the
beginning of a long prompt, which made it harder to see if the model was
generating new tokens or just repeating the prompt. This PR updates the
function so that both the beginning and end of a long prompt are
printed.

MAX_INTEGRATION_TESTS_REV_ID: 6293b3f028899d7327f2da9d4e201152fd2aeffc