[Models] Print both beginning and end of long prompt for MAX models (#65181)
This is a follow-up to PR \#64985, which updated the printing for models
running on Torch. Here, we copy the behavior for models running on MAX.
Having both the beginning and the end of a long prompt printed on screen
makes it easier to see at a glance if the model is generating new tokens
or simply repeating the prompt.

MAX_INTEGRATION_TESTS_REV_ID: 6fc0b18f6b8d6c57cad72de1dd8a402c3a2798d7