[KGEN] Actually insert the apply to the beginning of function if body parent is a Generator. (#65086)
We got something like 👇 in IR
```
 hlcf.if %7 {
      // define in if
      kgen.param.apply *"(lifted)apply_2" = ...
 } else {
      // used in else
      kgen.param.assert <*"(lifted)apply_2">
 }
```
because LiftAndFoldApply insert `apply` as near where the original op is
but `hlcf.if` is not a param decl interface (hence not a parameter
domain). This works with current elaborator because we reorder
`kgen.param.apply` based on parameter decl domain, but this is
problematic for interpreting parametric functions where we don't reorder
ops and the use in the `else` region will miss the parameter value
because the `then` region is not interpreted.

In `LiftAndFoldApply` there is a comment that

```
// If the parent is a function, extract 'apply' operators and place them
// at the start of the body.
```
But the source code doesn't actually do that :-(

- [x] Actually insert the `apply` to the beginning of generator if body
parent is a GeneratorOp.
- [x] Add unit test.

MODULAR_ORIG_COMMIT_REV_ID: d7cfb560a21ca65b5a7d6a18e153683d437b8e77