[bazel] Sync internal and external version
The more we push to the open source repo the harder it is to avoid
features we're relying on from this newer version. We can attempt to
avoid those features if a compelling case comes up.

MODULAR_ORIG_COMMIT_REV_ID: 4feb8eb05be2280bdc0955cd15423da5558a1d1d