[External] [docs] EqualityComparable example incorrect for `__ne__` (#65224)
[External] [docs] EqualityComparable example incorrect for `__ne__`

instances and unequal if either the "re" part OR (not AND) the "im" part
are unequal

<!--
Thanks for submitting a pull request,
your contribution is really appreciated!

If possible, add a link to the issue you are
trying to solve in the pull request description.

If your pull request is big (> 100 lines), consider splitting it
into multiple pull requests as it may accelerate the review process.
-->

Co-authored-by: Julian <juliansteden@gmail.com>
Closes modular/modular#4946
MODULAR_ORIG_COMMIT_REV_ID: 794258e5cc4238b367d0baabd33633f21c9c2279