[Kernels] Relax BN Constraint for Enabling Shared Memory Packing in Matmul
The previous power-of-two requirement for shared memory packing in the
H100 matmul kernel has been removed. The new requirement is that BN must
be a multiple of 8, which aligns with the existing constraint of the
wgmma instructions.

MODULAR_ORIG_COMMIT_REV_ID: 062e9d5e165c1114c31c6c102e515394bc5af25e