[bazel] Globally apply sitecustomize
Remove the `use_sitecustomize` flag and default it to enabled
everywhere. This script normalizes the environment to python code ignore
Bazel specifics. This is important now that the mojo importer is in Max
as it requires all the Mojo env variables to be correct.

MODULAR_ORIG_COMMIT_REV_ID: a6940b467c46b28741fc0f93902241f3f5ea5a51