[Interfaces] Replace STOP_STREAM with EngineResult
STOP_STREAM is a singleton, which cannot be deserialized appropriately
with Msgpack as a result, a new class EngineResult has been introduced.
This class identifies if the result is either cancelled or completed,
along with providing the appropriate result (if provided). This makes
the response queue serializable, and the system more explicit.

Along with this, a series of doc strings have been added to the
EngineQueue, which provide some useful context, on how these classes
work and why they are structured the way they are.

MAX_PYTHON_ORIG_REV_ID: b34c5e0b3e35cf1dfd52c111ead1451ab69e5dee