[LSP] Don't generate duplicate progress tokens (#65411)
Fixes MOTO-1192, caused by failing to seed the RNG we used to generate
progress tokens. This caused us to generate duplicate progress tokens,
which breaks the protocol. A flaw in the LSP proxy amplified this issue
into an outright crash of the LSP proxy in VS Code.

This PR fixes both of these issues. In the proxy, we now properly pass
the error reported by the client back up to the server. On the server,
we correctly seed the RNG.

I've added a test for this behavior that should prevent regressions of
the server-side issue. Testing the proxy is effectively impossible right
now, though.

MODULAR_ORIG_COMMIT_REV_ID: 61b653cb59ef9e6d3c827592b0411040a629b4a4