[KGEN] More fix for LiftAndFlodApply (#65639)
Fixing another issue with LiftAndFoldApply that we got something like 👇
in IR
```

 kgen.param.if<..> {
   hlcf.if %7 {
      // define in if
      kgen.param.apply *"(lifted)apply_2" = ...
   } else {
      // used in else
      kgen.param.assert <*"(lifted)apply_2">
   }
}
```

Previous fix in [Internal link] lifts
`(lifted)apply_2` to the proper scope if `hlcf.if` is not inside a
`DeclInterface` op (e.g. `kgen.param.if`). For the case above, we want
the apply to be lifted to the beginning of the `kgen.param.if`'s region
where the `hlcf.if` is in, instead of inserted close to where the
original op is.

- [x] Fix apply insertion point when it is inside of a DeclInterface op.
- [x] Add unit test.

MODULAR_ORIG_COMMIT_REV_ID: 3a901753e66269832c62fa65c881cdaded7bb166