[KGEN] Extend CallOpInterface to extract symbol reference (#65544)
Add a getCalleeSymbol method to CallOpInterface. This method will
attempt to extract a SymbolRefAttr from the result of getCallee.

The structure allows us to simplify the code in CallGraphUtils to
support operations which do not carry a SymbolConstantAttr, which is
common in Graph Compiler code.

MODULAR_ORIG_COMMIT_REV_ID: 335b0e43d32929e4fbd16ea9f847572dba67cad4