[CI] Disable --test_keep_going for GPUs
This makes bazel stop as soon a single test fails. This is worse UX
since it's nicer to see all the tests that you potentially broke at
once, but it's better for capacity issues so that a failing job doesn't
consume all the GPU resources when we know it will be red.

MODULAR_ORIG_COMMIT_REV_ID: 33dd0f518db6f7de7634edf5047afb517db57794