[Mojo] Expose Endianess  (#65771)
In order to test the folding of SIMDAttrs dependent on the endianess of
the target, we need a way to specify the endianess in the module. This
PR adds a field to the DataLayout attribute to store this value. It also
enables you to then use this value by grabbing the target info from the
module op instead of from the initialize function, which would always
overwrite the target with the host machine info, thus overwriting the
target info specified in the test file.

MODULAR_ORIG_COMMIT_REV_ID: 8147922f7f55433fc4181ab4df2a92e44bca3b01