[stdlib] polish: Rename `List.append(Span)` to `extend()`
This renames an overload of `append()` that took a `Span` to instead be
another overload of `extend()`.

The difference between `append()` and `extend()` in Python (and in Rust)
is that the former grows the length of the underlying List by exactly 1,
whereas the latter can grow the list between 0 and an arbitrary number
of elements based on the size of the collection passed as an argument.

Being able to reason about the difference in growth semantics at a
glance is an important readability property.

MODULAR_ORIG_COMMIT_REV_ID: db14dce0bd43a2ce870cb8e1f538a802155e92e1