[KVCache] Move external_claim and release to use RequestID not seq_id
This also makes a series of changes in the Block Manager and Paged Cache
to ensure that the cache_seq_id is not used to track state. In Fetch
scenarios in which block ids are used to retrieve memory sections, the
cache seq id is retrieved from internal state and not the
InputContext.cache_seq_id.

MODULAR_ORIG_COMMIT_REV_ID: 394146ec415a3ed478825e2340c3bfe107fb8620