[SDK] Add gather bounds checking on CPU and assert on GPU
When the gather operation runs on CPU, it will check for out-of-bounds
accesses and provide a useful error to the user. On GPU it is more
difficult to check and error properly, so the implementation uses
`debug_assert` when an out of bounds condition is detected to avoid
accessing memory incorrectly.

MODULAR_ORIG_COMMIT_REV_ID: ec3878132c011f61b1b16eee68f68e54ff90fd27

Signed-off-by: Amet13 <admin@amet13.name>