[UCX] Build UCX with verbs
Building UCX with verbs is needed for fast inter-node transfers with
RDMA.

This PR serves only to build UCX with verbs and allow NIXL to load the
UCX plugin. It does NOT enable fast inter-node transfers. This is
prerequisite
work for that.

Building UCX with verbs creates uct_ib and uct_ib_mlx5 transports. These
transports depend on `libibverbs.so.1` and `libmlx5.so`. As such, we
vendor
two different versions of UCX plugin. One without verbs and one with
verbs.
We decide which to load based on the presence of these system libraries.

Note that we do NOT add another tarball to bazel! We simply update
ucx_prebuilt.BUILD to create two `:ucx` targets based on which UCT
backends
are statically linked in.

MODULAR_ORIG_COMMIT_REV_ID: c0ef3c62853d853849070be89c179751794f0541