[LSP] Enable parallel NodeJS test execution (#66014)
As the title says, enables parallel test execution using Mocha's
parallelization support. This required some extra work in the reporting
end of things; the junit reporter and/or the multi-reporter wrapper we
were using previously seem to be broken with parallelization, so I just
replaced them with a simple wrapper that uses the stock Mocha reporters.
The XUnit reporter that Mocha ships with seems to provide more
information to Bazel (specifically about skipped tests), so this is a
win-win-win since it lets us drop the dependencies.

As part of merging this PR, I fixed MOTO-1268. This was caused by
occasionally (due to AsyncRT threading) invoking a pure virtual method
in the constructor of `MojoDocument`.

MODULAR_ORIG_COMMIT_REV_ID: dc107f4f21c3f19e8091a65da6617e90b29d5127