[mojo-parser] Fix issues in lazy resolution of nested trait decls (#66643)
Fixes a couple bugs that existed with the recent changes to make
resolution of trait methods and aliases lazier.

In the case of trait alias decls the issue was an alias with the same
name showing up in multiple parent of a trait while the trait itself did
not define the alias.

In these cases a decl would be created in the child trait for the first
parent trait alias encountered then upon encountering the second
parent's trait alias logic in trait body resolution would erroneously
assume that the child trait had an override and improperly set an
attribute on the alias decl it had created in the previous step.

In the case of trait method decls the issue occurred because I had
missed a spot in overload resolution that should have had logic to skip
over lit.fn ops that were marked as 'disabled'.

MODULAR_ORIG_COMMIT_REV_ID: 7533024f2cd2523a61d699f033624b5deeeff347