[DeepseekV2] Plumb additional kv cache params fields
Previously, the `--enable-prefix-caching` flag was ignored for
DeepseekV2
and we defaulted to `False`.

It is a bit suspect that we take a `KVCacheConfig` and use it to create
a new `KVCacheConfig` object. It is very easy to forget fields with this
pattern.

MODULAR_ORIG_COMMIT_REV_ID: 0801fc3cb5f0d0657a50e959a429c4400fcf743b