[mojo] Correctly set --target-features for cross compilation (#66817)
Previously even if the user passed `--target-triple` and `--target-cpu`,
the CPU features that ended up being used were that of the host, making
it difficult to cross compile. Now this fixes that single case and
fetches the features from the triple / CPU if they are not provided.

This still leaves a TODO that if you only pass `--target-triple`, mojo
still defaults the CPU to the host, and ideally it could determine a
default that was valid for the triple instead.

There is another notable hole if you go through the mcpu/march codepath
the OS defaults to the host OS.

This fixes compiling mojo code on linux that is then run on macOS.

MODULAR_ORIG_COMMIT_REV_ID: dfe596eef618f25f49b426ee527b0083a493f05b