[mojo-lang] Rework some of struct emission logic. (#66846)
This reworks struct emission logic to follow the parser more
naturally: instead of emitting a return into the end of a del and
copyinit/moveinit and munging it, we now create an empty function
and fill it in during body resolution / population.  This also uses
`emitNormalReturn` instead of inlining it, to reduce duplication and
sources of truth.

MODULAR_ORIG_COMMIT_REV_ID: 039eb6f8e6beb6e85da2107c52761152551effda