[mojo-lang] Implement the `deinit` argument convention.
This implements the `deinit` argument convention, adding parsing
and semantic analysis support.  This updates just a couple of places
to use it to test it, but does not do a mass migration.

This also expands the proposal text a bit to be more useful.

MODULAR_ORIG_COMMIT_REV_ID: 72680df1b9df1190214436885f3806c652fa32a2