[Serve] Proper handling of sampling defaults
Type checking of the OpenAI schema is disabled for... historical
reasons, we'll say. So some invalid types have snuck in. In particular,
SamplingParams shouldn't have any `None`s in it, but we stick some Nones
in it that MyPy doesn't detect. These should probably be defaulted
rather than making their way in as an unintended type.

I'm not terribly pleased about having to repeat the definition of most
fields of SamplingParams, but I wasn't able to think of another way
(that doesn't destroy type checking in other ways), and I asked Cursor
and it came up with an even worse way of doing it, so this is all I got.

MAX_PYTHON_ORIG_REV_ID: 5b29bb4d0c6447b42f7a0f609d9eee7932f98415