[Graph] Fix signature of ops.while_loop
We have always supported returning a single value from the body
function, but the signature we declare and the documentation
contradicted that, even though a very example in the documentation and
many of the uses (including in the tests) return a single value. This
wasn't being caught because most of the uses didn't have type
annotations on the functions. Annotate those and fix the signature and
documentation, so the implemented functionality (and uses) match the
signature and documentation.

MODULAR_ORIG_COMMIT_REV_ID: 08e8a5186de5697dd801880a7cc29d7987731eb1