[Pipelines] Add flag to configure buffer cache
Exposes an experimental flag to set the buffer cache size via the CLI.
This just sets an env var under the hood. We prefix with experimental
because (1) this is new and (2) we should eventually fold this in with
the `--device-memory-utilization` flag that dictates the amount of
memory we can allocate for the kvcache.

MODULAR_ORIG_COMMIT_REV_ID: ce7018f836e5201e0d007772772e1af303f2bbbc