[mojo] Fix another cross compilation OS issue (#67272)
Previously if you were going through the march codepath for determining
the target info, the OS would be set to the current host OS no matter
what. Now we default that to the passed target triple, which itself
defaults to the host OS, which allows overriding that value.

As part of this we no longer need to wipe out the arch in the aarch64
case, which otherwise results in an invalid triple in the end.

MODULAR_ORIG_COMMIT_REV_ID: ae93b3b07cb50f660216c7159fd2cfd3be71f535