[mojo-lang] Revise discard pattern emission to not load lvalues. (#67393)
### [mojo-lang] Revise discard pattern emission to not load lvalues.

This changes discard pattern emission `_ = <expr>` to only load the
RHS if it is a computed LValue like a getitem call, but otherwise
not to call the copy constructor.  This allows the discard pattern
to work naturally with non-copyable types and is somewhat more
efficient in theory.  In practice, this hopefully resolves
confusion around this behavior, which keeps coming up.

MODULAR_ORIG_COMMIT_REV_ID: a9e09d0fb8ff11937037ffc2552d40054b7d7d27