[DI] Use `text_batch_constructor` for `decode_scheduler`
This extends the `decode_scheduler` to support request preemptions. This
is needed to not fail when `max-batch-size` is larger.

One request was preempted and the benchmark completed without issue.

Note that the strategy we adopt for reuqest preemption is just to let
the
decode scheduler run the CE again. As such, the decode scheduler has to
deal with chunked prefill and such. We do not resend the CE request
backend
to prefill worker. I am scared that this alternative strategy may get
into
a loop if implemented poorly.

Note that the `text_batch_constructor` class allows us to reuse a bunch
of
code. However, the abstraction is not fully baked yet. More cleanups to
come!

```
22:45:25.218 INFO: 578063 MainThread: max.serve: Executed TG batch with 8 reqs | Terminated: 0 reqs, Pending: 0 reqs | Target: 8/INF toks | Prompt Tput: 36.2 tok/s, Generation Tput: 362.2 tok/s | Batch creation: 31.64us, Execution: 220.84ms | KVCache usage: 98.1% of 3583 blocks, Cache hit rate: 0.0% | All Preemptions: 0 reqs
22:45:25.440 INFO: 578063 MainThread: max.serve: Executed TG batch with 8 reqs | Terminated: 1 reqs, Pending: 0 reqs | Target: 8/INF toks | Prompt Tput: 36.2 tok/s, Generation Tput: 361.9 tok/s | Batch creation: 32.57us, Execution: 221.08ms | KVCache usage: 97.4% of 3583 blocks, Cache hit rate: 0.0% | All Preemptions: 0 reqs
22:45:25.441 INFO: 578063 MainThread: max.serve: Preempted a request due to lack of KV pages. This can affect the end-to-end performance. Consider increasing device-memory-utilization to provide more KV cache memory. Total preemption count: 1.
22:45:25.478 INFO: 578035 Thread-2 (_listen_loop): max.serve: received request from decode node.
22:45:25.560 INFO: 578035 MainThread: max.serve: initiating transfer from prefill worker.
22:45:25.560 INFO: 578035 MainThread: max.serve: initiating transfer from prefill worker.
22:45:25.561 INFO: 578035 MainThread: max.serve: initiating transfer from prefill worker.
22:45:25.561 INFO: 578035 MainThread: max.serve: Executed CE batch with 4 reqs | Terminated: 0 reqs, Pending: 54 reqs | Target: 8192/8192 toks | Prompt Tput: 13.8K tok/s, Generation Tput: 6.7 tok/s | Batch creation: 205.38us, Execution: 594.62ms | KVCache usage: 5.4% of 3583 blocks, Cache hit rate: 0.0% | All Preemptions: 0 reqs
```
MAX_TESTS_TESTS_ORIG_REV_ID: ad1aa2ad21c571c2fd6c1e86288733824bc33bbf