[Context] Fix how we determine if a context needs Context Encoding
In the scheduler, we return a context to the ce queue for additional ce
by querying the `is_ce` property. This previosuly just checked if the
active
length was greater than 1. However, there are situations where the
active length
is greater than 1, but the context is not in CE mode. One instance is
for
spec decoding, where we create a bunch of draft tokens we verify.

Now the `needs_ce` property checks if the start index is less than the
prompt len.
If the start idx >= prompt len, then we know that all prompt tokens have
been
encoded.

MAX_INTEGRATION_TESTS_REV_ID: 02cbd59604ce6379985957cfd6fd58275a9a0570