[NIXL] Make test_notification_latency test better and add docstring to ucx_backend.cpp
The test was passing even though the progress thread is not sending out
notifs immediately (I test this by commenting out the code). This is
because
the transfer was so small that it finished before `postXfer` returned.
If that happens, the notif is sent out even without the sender polling
with `checkXfer`. Now the test uses a larger buffer so that it completes
AFTER the `postXfer` call returns. With this change, the test fails if
I now comment out the notification-sending code in the progress thread
and can appropriately catch regressions.

I also address Brendan's feedback from the previous PR and add some
brief
comment about locking in `ucx_backend.cpp`.

MAX_INTEGRATION_TESTS_REV_ID: 4ced4a6da00a1e96938332b0e17a2961ebbf298e