[stdlib] Add tests for `T.__*__is_trivial`.
An type is given as an parameters and the struct is instantiated.
The struct value is then interacted with in any ways.

As interactions goes, the struct register the corresponding events for:
- `__init__`
- `__moveinit__`
- `__copyinit__`
- `__del__`

The event registered depends of `T`.
For example, when `T.__moveinit__is_trivial`,
The event registered on `__moveinit__` is `event_move | event_trivial`.

We can then compare the sequence of events to the expected one.