[stdlib][kernels] Update the multimem functions
For the functions multimem_ld_reduce and multimem_st:

The parameter `count` is the vector size with allowed values 1, 2, 4, or
8. This was incorrectly defined as the number of GPUs.

The parameter `output_type` is the width of the type. It's not in
general the simd_width. The range depends on the type and can be 1, 2,
or 4.

Define `bits = count * output_width * sizeof[dtype]() * 8`. The
requirement is that bits has to be 32, 64, or 128.
We can then define  `simd_width = count * output_width`.

https://docs.nvidia.com/cuda/parallel-thread-execution/#multimem-addresses
Here is an image showing the allowed types and for a given vector size.
For example `count = 4`, `output_width = 2` and type `bfloat16` maps to
`.v4` and `.bf16x2` in the PTX.
<img width="2519" height="574" alt="image"
src="https://github.com/user-attachments/assets/79d0c8b4-973b-4271-ba06-9109779c240d"
/>

In general it's best to choose an `output_width` so that type uses a
32-bit register (except for fp64). For example, if the type is
`bfloat16` then `output_width` should be 2 so that it packs two
`bfloat16` values (called `bf16x2` in the PTX) in one 32-bit register.

An simpler `multimem_ld_reduce` function has been added that instead of
having count and output_width as parameter only the desired simd_width
is a parameter. It defines the output_width and count as
```
    alias output_width = 4 // sizeof[dtype]()
    alias count = simd_width // output_width
```

This is the public portion of my message.

MODULAR_ORIG_COMMIT_REV_ID: 9f6c972d4866644a5f8dfc5fd8c7d245fb811250