[SDK] Add some more development lookup logic
This allows us to support a new development case where we're in a
development venv that doesn't have the full wheel installed, but has all
of the transitive things it needs in the standard wheel-like paths.

Theoretically we could skip this with MODULAR_DERIVED_PATH but there are
some cases where mojopkgs are required over the default source based
mojo lookup path. In that case we have to have the built mojopkgs that
are now included in venvs.

MODULAR_ORIG_COMMIT_REV_ID: dc550882a58aab6787851b246092408429d45285