[Kernels] Improve location information for errors
The calls to `__call_location` in the `_checked` helper will always
print the call location of the `_checked` function, which is unlikely to
be useful in the various `DeviceContext` functions. The locations will
tend to point to `DeviceContext` internals rather than the callsite of
the offending `DeviceContext` method.

This change tries to capture the location at the correct point for
simple methods like `synchronize`.

MODULAR_ORIG_COMMIT_REV_ID: a3c7fe635867e5ea11115e54b8f401b1da726d76