[kgen] flatten GeneratorAttr repr (#67923)
Instead of storing a GeneratorType inside GeneratorAttr, do away with
this additional level of nesting and store the input param types and
metadata directly on the GeneratorAttr. This encoding ensures the body
value and body type of the GeneratorAttr are always at the same level of
nesting within an attr/type tree. If we instead stored a GeneratorType
in this attribute, the body of the GeneratorType would be at a deeper
level of nesting than the body of the GeneratorAttr, leading to
inconsistencies.

The usefulness of this will be tested here
[Internal link]

MAX_PYTHON_ORIG_REV_ID: a170712314bbf64ba7346a9aba57e71341142c06