[stdlib] Dict: re-use removed slots instead of always using empty slots.
**Overview**

If we don't re-use the removed slots, the next empty slots gets further and further.

And we need to encounter one empty slot to determine when the dict don't contains an key.
We also currently need to encouter an empty slot to insert an key_value entry.

Some empty slots are on the path of many hashes.
This can compound into an huge slowdown/issues so this patch fix that.
The result is also an massive speedup that users can naturally appreciate.

**Why it works**

While in the loop,
we place the first encoutered removed slot in an optional.
When we reach the empty slot, it means the dict don't contain the key.
Instead of using the empty slot, we use the removed one if any was encountered on the way!

That entry lookup will be faster because it won't have to loop as far as the empty slot was.
The next insertion that collide with this entry won't need to go further than the empty slot.
Because if we'd use that empty slot before, the new insertion would need to loop at least that amount + 1.

So this prevent the empty slots to drift further and further away.
(fixing an issue and creating an speedup by removing an slowdown)

**While in what loop ?**

The dict currently takes an hash and trucate it by modulo capacity (hash%256 for example).
So if the hash is 1, the first slot checked is 1 (1%256).
If it needs to check next slot: multiply and truncate again.
What happens after a lot of "check next", is an wrap around: (258%256) = 2
And this happen until we get either an existing slot for the key or an empty slot.

So, the amount of any looping (time) depends on:
- how far the empty slot is. (don't contain the key, this is an insertion)
- how far the slot for an entry is. (contains the key, this is an lookup)

What we need is to minimize how far the first empty slot is (by not using it)!
by reusing removed slots, we also minimize how far the new entry is.

**+Speedup: useful**

And what we get is faster lookup and insertions, sometimes massive,
but only in situations where we used pop.
because pop turns an used slot into an removed slot.

**-Slowdown: security**

This solves an problem with pop.
The next empty slot don't drift further an further away anymore.
For large dicts, it was an larger problem.
This is the main reason we should re-use the removed slots

**Conclusion**

While not being an expert at all,
i think it make sense to re-use the removed slots.
The pr description shows that this logically can work.
(especially for the goal of preventing large slowdowns)

@JoeLoser asked for rebase to merge, this is a new PR (merge conflict).
Please do as always, all you can to make sure this is an good change !
I really don't wanna have the working dict having an bug.
Users naturally love this type and often benchmark it.