[mojo-parser] Fix alias access on type parameter (#67946)
Fixes MOCO-2356.

The main problem is accessing an associated alias on an instance of a
parametric type. We know statically that the type conforms to some
trait, but not the exact type. In cases like this, we need to
automatically lift the access from being on the instance to being on the
trait.

MODULAR_ORIG_COMMIT_REV_ID: 5108ae5789e92ec68c469fbbb9f6032e5e714aae