[mojo] Fix overload resolution bug with defaulted trait methods (#68007)
If a defaulted trait method was overridden in a struct the corresponding
ASTDecl would be marked as 'disabled'. This proved problematic in the
case of packages as there would be no way to distinguish if a particular
defaulted trait method was defaulted or not after round tripping it from
bytecode.

Fix this by making 'disabled' an attribute of lit.fn ops instead
ensuring it's survives the roundtrip.

MODULAR_ORIG_COMMIT_REV_ID: 6fa8b308881a3c94a7314ea10535f147337539e1