[LSP] Fix LSP server crash due to use-after-free (#68141)
Progress reporting in the Mojo LSP causes the server to crash frequently
on Linux. The root cause is a use-after-free bug where a MojoDocument is
accessed after it was freed. The closure given to
`ProgressManager::withProgress` was not properly extending the lifetime
of the MojoDocument, which is managed as an AsyncValue through
a MojoDocumentRef.

The fix increments the reference count of the MojoDocument through
a MojoDocumentRef and ties that to the lifetime of the withProgress
closure.

MODULAR_ORIG_COMMIT_REV_ID: 21a517f8e57d747a677992cd307d7caaac567947