[stdlib] Add tests for struct inheriting triviality (T.__*__is_trivial).
An type with an parameter T is instantiated.
It sets its own triviality based on T.

For example,
if T is Float64,
Self.__copyinit__is_trivial is set to T.__copyinit__is_trivial.

Then the type goes trough an triviality tester,
to determine if it works as expected.

---

Possible auto-optimizations related to inherited triviality:
- `InlineArray[32, InlineArray[32 Int]]` -> Fully trivial
- `Variant[Int64, Bool]` -> Fully trivial
- `List[Optional[Int]]` -> No hint_trivial_type