[mojodoc] Add MojoDoc support for __init__.mojo files with APIs
If the `__init__.mojo` file included APIs, the MojoDoc templates would
add `__init__` to the pathname, which is not what we want. The
functions/structs/traits defined in `__init__.mojo` should create pages
inside the current directory, not a subdirectory like we do for other
`.mojo` module files.

Funny thing is, it appeared that this was causing a docsite breakage,
but the real issue was that the new `iter` module did not include a
description so it was falling back to the default description from the
Jinja template, which uses back tics for code font and that’s what was
breaking the listing card widget. So this also removes that code font. 🤦

Also add function docstrings for posterity.

MODULAR_ORIG_COMMIT_REV_ID: ad725952f302288fa587d9418112b8d4b890fb08